Changes in line of request
