+++ test/config.dest.json
@@ -5,7 +5,7 @@
     "baseurl": "https://amusementclub.nyc3.digitaloceanspaces.com",
     "shorturl": "https://amuse.noxc.dev",
     "daily": 72000000,
-    "audit": {
+    "auditc": {
         "channel": "DISCORD_VALID_CHANNEL_ID_GOES_HERE"
     }
 }
